+++ src/tests/ascent/t_ascent_transform.cpp
@@ -545,6 +545,172 @@ TEST(ascent_translate, test_bad_params)
     }
 }
 
+//-----------------------------------------------------------------------------
+TEST(ascent_translate, test_reflect_x)
+{
+    if(!vtkm_avalible())
+    {
+        return;
+    }
+
+    std::string output_file;
+    conduit::Node data;
+    setup("tout_transform_reflect_x",data,output_file);
+
+    conduit::Node actions;
+    conduit::Node &add_pipelines = actions.append();
+    add_pipelines["action"] = "add_pipelines";
+    conduit::Node &pipelines = add_pipelines["pipelines"];
+
+    pipelines["pl0/f1/type"] = "transform";
+    pipelines["pl0/f1/params/translate/x"]= 10.0;
+    pipelines["pl0/f1/params/translate/y"]= 10.0;
+
+    // filter knobs
+    pipelines["pl1/f1/type"] = "transform";
+    pipelines["pl1/f1/params/reflect/x"]= 1.0;
+    pipelines["pl1/pipeline"] = "pl0";
+
+    conduit::Node &add_scenes = actions.append();
+    add_scenes["action"] = "add_scenes";
+    conduit::Node &scenes = add_scenes["scenes"];
+
+    // render transformed
+    scenes["s1/plots/p1/type"]  = "pseudocolor";
+    scenes["s1/plots/p1/field"] = "braid";
+    scenes["s1/plots/p1/pipeline"] = "pl1";
+    // and orig
+    scenes["s1/plots/p2/type"]  = "pseudocolor";
+    scenes["s1/plots/p2/field"] = "braid";
+    scenes["s1/plots/p2/pipeline"] = "pl0";
+
+    scenes["s1/image_prefix"] = output_file;
+
+    std::cout << actions.to_yaml() << std::endl;
+
+    Ascent ascent;
+    ascent.open();
+    ascent.publish(data);
+    ascent.execute(actions);
+    ascent.close();
+
+    // check that we created an image
+    EXPECT_TRUE(check_test_image(output_file));
+    std::string msg = "An example transform filter using reflect across x axis.";
+    ASCENT_ACTIONS_DUMP(actions,output_file,msg);
+}
+
+//-----------------------------------------------------------------------------
+TEST(ascent_translate, test_reflect_arb)
+{
+    if(!vtkm_avalible())
+    {
+        return;
+    }
+
+    std::string output_file;
+    conduit::Node data;
+    setup("tout_transform_reflect_arb",data,output_file);
+
+    conduit::Node actions;
+    conduit::Node &add_pipelines = actions.append();
+    add_pipelines["action"] = "add_pipelines";
+    conduit::Node &pipelines = add_pipelines["pipelines"];
+
+    pipelines["pl0/f1/type"] = "transform";
+    pipelines["pl0/f1/params/translate/x"]= 10.0;
+    pipelines["pl0/f1/params/translate/y"]= 10.0;
+
+    // filter knobs
+    pipelines["pl1/f1/type"] = "transform";
+    pipelines["pl1/f1/params/reflect/x"]= 1.0;
+    pipelines["pl1/f1/params/reflect/y"]= 1.0;
+    pipelines["pl1/pipeline"] = "pl0";
+
+    conduit::Node &add_scenes = actions.append();
+    add_scenes["action"] = "add_scenes";
+    conduit::Node &scenes = add_scenes["scenes"];
+
+    // render transformed
+    scenes["s1/plots/p1/type"]  = "pseudocolor";
+    scenes["s1/plots/p1/field"] = "braid";
+    scenes["s1/plots/p1/pipeline"] = "pl1";
+    // and orig
+    scenes["s1/plots/p2/type"]  = "pseudocolor";
+    scenes["s1/plots/p2/field"] = "braid";
+    scenes["s1/plots/p2/pipeline"] = "pl0";
+
+    scenes["s1/image_prefix"] = output_file;
+
+    std::cout << actions.to_yaml() << std::endl;
+
+    Ascent ascent;
+    ascent.open();
+    ascent.publish(data);
+    ascent.execute(actions);
+    ascent.close();
+
+    // check that we created an image
+    EXPECT_TRUE(check_test_image(output_file));
+    std::string msg = "An example transform filter using reflect across arbitrary axis.";
+    ASCENT_ACTIONS_DUMP(actions,output_file,msg);
+}
+
+
+//-----------------------------------------------------------------------------
+TEST(ascent_translate, test_reflect_y)
+{
+    if(!vtkm_avalible())
+    {
+        return;
+    }
+
+    std::string output_file;
+    conduit::Node data;
+    setup("tout_transform_reflect_y",data,output_file);
+
+    conduit::Node actions;
+    conduit::Node &add_pipelines = actions.append();
+    add_pipelines["action"] = "add_pipelines";
+    conduit::Node &pipelines = add_pipelines["pipelines"];
+
+    pipelines["pl0/f1/type"] = "transform";
+    pipelines["pl0/f1/params/translate/x"]= 10.0;
+    pipelines["pl0/f1/params/translate/y"]= 10.0;
+
+    // filter knobs
+    pipelines["pl1/f1/type"] = "transform";
+    pipelines["pl1/f1/params/reflect/y"]= 1.0;
+    pipelines["pl1/pipeline"] = "pl0";
+
+    conduit::Node &add_scenes = actions.append();
+    add_scenes["action"] = "add_scenes";
+    conduit::Node &scenes = add_scenes["scenes"];
+
+    // render transformed
+    scenes["s1/plots/p1/type"]  = "pseudocolor";
+    scenes["s1/plots/p1/field"] = "braid";
+    scenes["s1/plots/p1/pipeline"] = "pl1";
+    // and orig
+    scenes["s1/plots/p2/type"]  = "pseudocolor";
+    scenes["s1/plots/p2/field"] = "braid";
+    scenes["s1/plots/p2/pipeline"] = "pl0";
+
+    scenes["s1/image_prefix"] = output_file;
+
+    std::cout << actions.to_yaml() << std::endl;
+
+    Ascent ascent;
+    ascent.open();
+    ascent.publish(data);
+    ascent.execute(actions);
+    ascent.close();
+
+    // check that we created an image
+    EXPECT_TRUE(check_test_image(output_file));
+    std::string msg = "An example transform filter using reflect across y axis.";
+    ASCENT_ACTIONS_DUMP(actions,output_file,msg);
+}
 
 //-----------------------------------------------------------------------------
 int main(int argc, char* argv[])